media-monitoring-knowledge-assitant/pyproject.toml
[project]
name = "media-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "fastapi[standard]>=0.116.1",
    "ipython>=9.4.0",
    "langchain[google-genai]>=0.3.27",
    "langchain-astradb>=0.6.1",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.6",
    "langsmith>=0.4.20",
    "llama-index>=0.13.3",
    "pylint>=3.3.8",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.7",
    "cassio>=0.1.10",
]
